refactor(src): add file extension on import paths
